기타 코드 모음/requirements.txt
google-cloud-bigquery  
pandas                 
db-dtypes              
pyarrow                
openpyxl